'Bad substitution' error from ${BASH_SOURCE[0]}
You will get a 'Bad substitution' error on the following 2 lines running the script
in Linux (e.g. Ubuntu 16.04), change /bin/sh on the first line to /bin/bash.
- scriptdir=$(cd $(dirname ${BASH_SOURCE[0]}); pwd)
- scriptname=$(basename ${BASH_SOURCE[0]} .sh)
This issue is because, bin/sh is a syslink to /bin/dash which does not support ${array[0]} 
There are several solutions, I have choosen to use bash explicitly.